Google code formatter

+++ src/org/thoughtcrime/securesms/jobs/AvatarDownloadJob.java
@@ -4,7 +4,10 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.support.annotation.NonNull;
 import android.util.Log;
-
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import javax.inject.Inject;
 import org.thoughtcrime.securesms.crypto.MasterSecret;
 import org.thoughtcrime.securesms.database.DatabaseFactory;
 import org.thoughtcrime.securesms.database.GroupDatabase;
@@ -24,12 +27,6 @@ import org.whispersystems.signalservice.api.SignalServiceMessageReceiver;
 import org.whispersystems.signalservice.api.messages.SignalServiceAttachmentPointer;
 import org.whispersystems.signalservice.api.push.exceptions.NonSuccessfulResponseCodeException;
 
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-
-import javax.inject.Inject;
-
 public class AvatarDownloadJob extends MasterSecretJob implements InjectableType {
 
   private static final int MAX_AVATAR_SIZE = 20 * 1024 * 1024;
@@ -42,11 +39,13 @@ public class AvatarDownloadJob extends MasterSecretJob implements InjectableType
   private final byte[] groupId;
 
   public AvatarDownloadJob(Context context, @NonNull byte[] groupId) {
-    super(context, JobParameters.newBuilder()
-                                .withRequirement(new MasterSecretRequirement(context))
-                                .withRequirement(new NetworkRequirement(context))
-                                .withPersistence()
-                                .create());
+    super(
+        context,
+        JobParameters.newBuilder()
+            .withRequirement(new MasterSecretRequirement(context))
+            .withRequirement(new NetworkRequirement(context))
+            .withPersistence()
+            .create());
 
     this.groupId = groupId;
   }
@@ -56,19 +55,19 @@ public class AvatarDownloadJob extends MasterSecretJob implements InjectableType
 
   @Override
   public void onRun(MasterSecret masterSecret) throws IOException {
-    String                encodeId   = GroupUtil.getEncodedId(groupId, false);
-    GroupDatabase         database   = DatabaseFactory.getGroupDatabase(context);
-    Optional<GroupRecord> record     = database.getGroup(encodeId);
-    File                  attachment = null;
+    String encodeId = GroupUtil.getEncodedId(groupId, false);
+    GroupDatabase database = DatabaseFactory.getGroupDatabase(context);
+    Optional<GroupRecord> record = database.getGroup(encodeId);
+    File attachment = null;
 
     try {
       if (record.isPresent()) {
-        long             avatarId    = record.get().getAvatarId();
-        String           contentType = record.get().getAvatarContentType();
-        byte[]           key         = record.get().getAvatarKey();
-        String           relay       = record.get().getRelay();
-        Optional<byte[]> digest      = Optional.fromNullable(record.get().getAvatarDigest());
-        Optional<String> fileName    = Optional.absent();
+        long avatarId = record.get().getAvatarId();
+        String contentType = record.get().getAvatarContentType();
+        byte[] key = record.get().getAvatarKey();
+        String relay = record.get().getRelay();
+        Optional<byte[]> digest = Optional.fromNullable(record.get().getAvatarDigest());
+        Optional<String> fileName = Optional.absent();
 
         if (avatarId == -1 || key == null) {
           return;
@@ -81,18 +80,33 @@ public class AvatarDownloadJob extends MasterSecretJob implements InjectableType
         attachment = File.createTempFile("avatar", "tmp", context.getCacheDir());
         attachment.deleteOnExit();
 
-        SignalServiceAttachmentPointer pointer     = new SignalServiceAttachmentPointer(avatarId, contentType, key, relay, Optional.of(0), Optional.absent(), 0, 0, digest, fileName, false);
-        InputStream                    inputStream = receiver.retrieveAttachment(pointer, attachment, MAX_AVATAR_SIZE);
-        Bitmap                         avatar      = BitmapUtil.createScaledBitmap(context, new AttachmentModel(attachment, key, 0, digest), 500, 500);
+        SignalServiceAttachmentPointer pointer =
+            new SignalServiceAttachmentPointer(
+                avatarId,
+                contentType,
+                key,
+                relay,
+                Optional.of(0),
+                Optional.absent(),
+                0,
+                0,
+                digest,
+                fileName,
+                false);
+        InputStream inputStream = receiver.retrieveAttachment(pointer, attachment, MAX_AVATAR_SIZE);
+        Bitmap avatar =
+            BitmapUtil.createScaledBitmap(
+                context, new AttachmentModel(attachment, key, 0, digest), 500, 500);
 
         database.updateAvatar(encodeId, avatar);
         inputStream.close();
       }
-    } catch (BitmapDecodingException | NonSuccessfulResponseCodeException | InvalidMessageException e) {
+    } catch (BitmapDecodingException
+        | NonSuccessfulResponseCodeException
+        | InvalidMessageException e) {
       Log.w(TAG, e);
     } finally {
-      if (attachment != null)
-        attachment.delete();
+      if (attachment != null) attachment.delete();
     }
   }
 
@@ -104,5 +118,4 @@ public class AvatarDownloadJob extends MasterSecretJob implements InjectableType
     if (exception instanceof IOException) return true;
     return false;
   }
-
 }
